+++ cmd/frontend/internal/cli/serve_cmd.go
@@ -80,7 +80,7 @@ func defaultExternalURL(nginxAddr, httpAddr string) *url.URL {
 }
 
 // Main is the main entrypoint for the frontend server program.
-func Main(githubWebhook http.Handler) error {
+func Main(githubWebhook, bitbucketServerWebhook http.Handler) error {
 	log.SetFlags(0)
 	log.SetPrefix("")
 
@@ -205,7 +205,7 @@ func Main(githubWebhook http.Handler) error {
 	}
 
 	// Create the external HTTP handler.
-	externalHandler, err := newExternalHTTPHandler(schema, githubWebhook, lsifServerProxy)
+	externalHandler, err := newExternalHTTPHandler(schema, githubWebhook, bitbucketServerWebhook, lsifServerProxy)
 	if err != nil {
 		return err
 	}