+++ web/src/enterprise/campaigns/detail/backend.ts
@@ -25,7 +25,7 @@ const campaignFragment = gql`
             username
             avatarURL
         }
-        changesetCreationStatus {
+        status {
             completedCount
             pendingCount
             state