+++ web/src/enterprise/repo/settings/backend.tsx
@@ -1,28 +1,43 @@
-import { Observable } from 'rxjs'
-import { map } from 'rxjs/operators'
-import { gql, dataOrThrowErrors } from '../../../../../shared/src/graphql/graphql'
 import * as GQL from '../../../../../shared/src/graphql/schema'
+import { dataOrThrowErrors, gql } from '../../../../../shared/src/graphql/graphql'
+import { map } from 'rxjs/operators'
+import { Observable } from 'rxjs'
 import { queryGraphQL } from '../../../backend/graphql'
 
 /**
- * Fetch LSIF dumps for a repository.
+ * Fetch LSIF uploads for a repository.
  */
-export function fetchLsifDumps({
+export function fetchLsifUploads({
     repository,
+    query,
+    state,
+    isLatestForRepo,
     first,
     after,
-    query,
-    isLatestForRepo,
-}: { repository: string } & GQL.ILsifDumpsOnRepositoryArguments): Observable<GQL.ILSIFDumpConnection> {
+}: { repository: string } & GQL.ILsifUploadsOnRepositoryArguments): Observable<GQL.ILSIFUploadConnection> {
     return queryGraphQL(
         gql`
-            query LsifDumps($repository: ID!, $first: Int, $after: String, $query: String, $isLatestForRepo: Boolean) {
+            query LsifUploads(
+                $repository: ID!
+                $state: LSIFUploadState
+                $isLatestForRepo: Boolean
+                $first: Int
+                $after: String
+                $query: String
+            ) {
                 node(id: $repository) {
                     __typename
                     ... on Repository {
-                        lsifDumps(first: $first, after: $after, query: $query, isLatestForRepo: $isLatestForRepo) {
+                        lsifUploads(
+                            query: $query
+                            state: $state
+                            isLatestForRepo: $isLatestForRepo
+                            first: $first
+                            after: $after
+                        ) {
                             nodes {
                                 id
+                                state
                                 projectRoot {
                                     commit {
                                         abbreviatedOID
@@ -33,7 +48,9 @@ export function fetchLsifDumps({
                                 inputRepoName
                                 inputCommit
                                 inputRoot
-                                processedAt
+                                uploadedAt
+                                startedAt
+                                finishedAt
                             }
 
                             totalCount
@@ -46,7 +63,7 @@ export function fetchLsifDumps({
                 }
             }
         `,
-        { repository, first, after, query, isLatestForRepo }
+        { repository, query, state, isLatestForRepo, first, after }
     ).pipe(
         map(dataOrThrowErrors),
         map(({ node }) => {
@@ -57,28 +74,31 @@ export function fetchLsifDumps({
                 throw new Error(`The given ID is a ${node.__typename}, not a Repository`)
             }
 
-            return node.lsifDumps
+            return node.lsifUploads
         })
     )
 }
 
 /**
- * Fetch LSIF upload with the given state.
+ * Fetch a single LSIF upload by id.
  */
-export function fetchLsifUploads({
-    state,
-    first,
-    query,
-}: GQL.ILsifUploadsOnQueryArguments): Observable<GQL.ILSIFUploadConnection> {
+export function fetchLsifUpload({ id }: { id: string }): Observable<GQL.ILSIFUpload | null> {
     return queryGraphQL(
         gql`
-            query LsifUploads($state: LSIFUploadState!, $first: Int, $query: String) {
-                lsifUploads(state: $state, first: $first, query: $query) {
-                    nodes {
+            query LsifUpload($id: ID!) {
+                node(id: $id) {
+                    __typename
+                    ... on LSIFUpload {
                         id
                         projectRoot {
                             commit {
+                                oid
                                 abbreviatedOID
+                                url
+                                repository {
+                                    name
+                                    url
+                                }
                             }
                             path
                             url
@@ -87,19 +107,28 @@ export function fetchLsifUploads({
                         inputCommit
                         inputRoot
                         state
+                        failure {
+                            summary
+                        }
                         uploadedAt
                         startedAt
                         finishedAt
                     }
-                    pageInfo {
-                        hasNextPage
-                    }
                 }
             }
         `,
-        { state: state.toUpperCase(), first, query }
+        { id }
     ).pipe(
         map(dataOrThrowErrors),
-        map(data => data.lsifUploads)
+        map(({ node }) => {
+            if (!node) {
+                return null
+            }
+            if (node.__typename !== 'LSIFUpload') {
+                throw new Error(`The given ID is a ${node.__typename}, not an LSIFUpload`)
+            }
+
+            return node
+        })
     )
 }